hmtl styling

+++ src/main/java/com/sparta/eng72/traineetracker/services/CodingGamesAPIService.java
@@ -104,7 +104,8 @@ public class CodingGamesAPIService implements CodingGamesAPI {
         return Double.parseDouble(df.format(score));
     }
 
-    public String getAssessmentGrade(int assessmentScore) {
+    public String getAssessmentGrade(double assessmentScore) {
+
         if(assessmentScore >= 95.8){
             return "A+";
         }else if(assessmentScore >= 91.6){
